moved modal to its own component
