+++ frontend/src/components/modal/modal.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { connect } from "react-redux";
 import { closeModal } from "../../actions/user_modal_actions";
-import  UserSettingsForm  from "./user_settings_container";
+import  UserSettingsForm  from "../user/user_settings_container";
 
 function Modal({ modal, closeModal }) {
   if (!modal) {
